+++ js_libraries/types/CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ## 3.4.3
 - Add `font-variation-settings`,`font-feature-settings` and `font-optical-sizing` CSS properties.
-- Add `experimental-recycle-available-item-before-layout` property for list.
+- Add `experimental-recycle-available-item-before-layout` and `enable-dynamic-span-count` property for list.
 
 ## 3.4.2
 - And `lynx.fetchBundle` api defines.